feat: Implement comment and review features with UI updates
- Added comment submission functionality with reply and quote options.
- Enhanced comment rendering with quoted comments and user info.
- Introduced a new CommentViewDrawer component for viewing comment details.
- Implemented review submission and rendering with user ratings.
- Updated MovieDetail.html to include comment and review sections.
- Created shared utility functions for API interactions and user data handling.
- Added CSS styles for comments and reviews to improve UI aesthetics.
- Refactored JavaScript files for better modularity and maintainability.
