+++ admin/src/components/Comments/CommentViewDrawer.module.css
@@ -1,0 +1,77 @@
+.drawerContainer {
+  width: 520px;
+  padding: 16px;
+  background-color: #1c1c1c;
+  color: #ffffff;
+}
+
+.headerText {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-weight: 500;
+  color: #ff9933;
+}
+
+.movieTitle {
+  font-size: 16px;
+  font-weight: 600;
+  margin-top: 4px;
+  color: #ffffff;
+}
+
+.quotedComment {
+  background: rgba(255, 255, 255, 0.05);
+  border-left: 3px solid #1890ff;
+  padding: 12px;
+  border-radius: 6px;
+  margin-top: 12px;
+  color: #ffffff;
+}
+
+.commentText {
+  font-size: 14px;
+  line-height: 1.5;
+  color: #ff9933;
+}
+
+.repliesContainer {
+  margin-top: 12px;
+  padding-left: 8px;
+  border-left: 2px dashed rgba(255, 255, 255, 0.2);
+}
+
+.replyText {
+  font-size: 13px;
+  line-height: 1.4;
+  color: rgba(255, 255, 255, 0.8);
+}
+
+.stats {
+  margin-top: 12px;
+  display: flex;
+  gap: 16px;
+  font-size: 14px;
+  color: #ff9933 !important;
+}
+
+.noData {
+  text-align: center;
+  margin-top: 48px;
+  color: rgba(255, 255, 255, 0.5);
+}
+
+.greyTextColor {
+  color: rgba(255, 255, 255, 0.6) !important;
+}
+.headerList {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.headerListItem {
+  margin-bottom: 4px;
+  color: rgba(255, 255, 255, 0.8);
+  font-size: 14px;
+}
