+++ admin/src/components/Reviews/index.tsx
@@ -1,8 +1,8 @@
 import { useState, useCallback, useEffect } from "react";
-import { Input, Select, Space } from "antd";
+import { Input, message, Modal, Select, Space } from "antd";
 import { SearchOutlined } from "@ant-design/icons";
 import SidebarLayout from "../layouts/SidebarLayout";
-import { fetchReviews, SortFields } from "../../api/reviewApi";
+import { fetchReviews, SortFields, DeleteReview } from "../../api/reviewApi";
 import styles from "./reviews.module.css";
 import useQuery from "../../hooks/useQuery";
 import { ReviewTable } from "./ReviewTable";
@@ -11,6 +11,8 @@ import { usePagination } from "../../hooks/usePagination";
 const sortOptions = Object.values(SortFields);
 
 const Reviews = () => {
+  const [Message] = message.useMessage();
+  const [ascending, setAscending] = useState(false);
   const [selectedSort, setSelectedSort] = useState<SortFields>(
     SortFields.CreatedAt
   );
@@ -20,15 +22,16 @@ const Reviews = () => {
     () =>
       fetchReviews({
         sortBy: selectedSort,
-        ascending: true,
+        ascending,
         page: 1,
         pageSize: 10,
       }),
-    [selectedSort]
+    [selectedSort, ascending, pagination.page, pagination.pageSize]
   );
 
   const {
     query: { status, response },
+    refetch,
   } = useQuery(fetchResponse);
 
   useEffect(() => {
@@ -36,6 +39,24 @@ const Reviews = () => {
       onTotalCountChange(response.totalCount);
     }
   }, [response]);
+
+  const handleDelete = async (reviewId: number) => {
+    Modal.confirm({
+      title: "Are you sure you want to delete this review?",
+      okText: "Yes",
+      okType: "danger",
+      cancelText: "No",
+      async onOk() {
+        try {
+          await DeleteReview(reviewId);
+          Message.success("Review deleted successfully");
+          refetch();
+        } catch (error) {
+          Message.error("Failed to delete review");
+        }
+      },
+    });
+  };
 
   return (
     <SidebarLayout>
@@ -54,12 +75,29 @@ const Reviews = () => {
             <div className={styles.sortLabel}>Sort by:</div>
             <Select
               value={selectedSort}
-              onChange={(value) => setSelectedSort(value)}
+              onSelect={(value) => {
+                if (value === selectedSort) {
+                  setAscending((prev) => !prev);
+                } else {
+                  setSelectedSort(value);
+                  setAscending(false);
+                }
+                onPageChange(1);
+              }}
               className={styles.select}
               variant="borderless"
               options={sortOptions.map((item) => ({
                 value: item,
-                label: <span>{item}</span>,
+                label: (
+                  <span>
+                    {item}
+                    {selectedSort === item && (
+                      <span style={{ marginLeft: 4 }}>
+                        {ascending ? "↑" : "↓"}
+                      </span>
+                    )}
+                  </span>
+                ),
               }))}
             />
           </div>
@@ -80,6 +118,7 @@ const Reviews = () => {
           currentPage={pagination.page}
           totalCount={pagination.totalCount}
           onPageChange={onPageChange}
+          onDelete={handleDelete}
         />
       </div>
     </SidebarLayout>
