+++ admin/src/components/Users/index.tsx
@@ -7,6 +7,7 @@ import { UserTable } from "./UserTable";
 import styles from "./Users.module.css";
 import useQuery from "../../hooks/useQuery";
 import { usePagination } from "../../hooks/usePagination";
+import { useSearchParams } from "react-router";
 
 const sortOptions = Object.values(SortFields);
 
@@ -14,6 +15,23 @@ const Users = () => {
   const [selectedSort, setSelectedSort] = useState<SortFields>(
     SortFields.CreatedAt
   );
+  const [ascending, setAscending] = useState(false);
+  const [searchParams] = useSearchParams();
+
+  useEffect(() => {
+    const sortParam = searchParams.get("sortBy") as SortFields | null;
+    if (
+      sortParam &&
+      Object.values(SortFields).includes(sortParam as SortFields)
+    ) {
+      setSelectedSort(sortParam as SortFields);
+    }
+
+    const pageParam = searchParams.get("page");
+    if (pageParam) {
+      onPageChange(Number(pageParam));
+    }
+  }, [searchParams]);
 
   const { pagination, onTotalCountChange, onPageChange } = usePagination(10);
 
@@ -21,11 +39,11 @@ const Users = () => {
     () =>
       fetchUsers({
         sortBy: selectedSort,
-        ascending: true,
+        ascending,
         page: pagination.page,
         pageSize: pagination.pageSize,
       }),
-    [selectedSort, pagination.page, pagination.pageSize]
+    [selectedSort, ascending, pagination.page, pagination.pageSize]
   );
 
   const {
@@ -67,12 +85,29 @@ const Users = () => {
             <div className={styles.sortLabel}>Sort by:</div>
             <Select
               value={selectedSort}
-              onChange={(value) => setSelectedSort(value)}
+              onSelect={(value) => {
+                if (value === selectedSort) {
+                  setAscending((prev) => !prev);
+                } else {
+                  setSelectedSort(value);
+                  setAscending(false);
+                }
+                onPageChange(1);
+              }}
               className={styles.select}
               variant="borderless"
               options={sortOptions.map((item) => ({
                 value: item,
-                label: <span>{item}</span>,
+                label: (
+                  <span>
+                    {item}
+                    {selectedSort === item && (
+                      <span style={{ marginLeft: 4 }}>
+                        {ascending ? "↑" : "↓"}
+                      </span>
+                    )}
+                  </span>
+                ),
               }))}
             />
           </div>
