+++ client/MovieDetail.html
@@ -123,6 +123,7 @@
         <div id="commentsSectionContent" class="tab-content active-content" data-tab-content="comments">
             <div class="comments-section"> <div id="commentsContainer">
                     </div>
+                    
                 <div class="comment-box">
                     <textarea placeholder="Add comment"></textarea>
                     <div class="btn btn-primary send-btn">SEND</div>
@@ -141,7 +142,7 @@
                     <div id="submitReviewBtn" class="btn btn-primary send-btn">Submit Review</div>
                 </div>
                 <div id="reviewsContainer">
-                    <!-- reviews get rendered here -->
+                    <p>No reviews yet.</p>
                 </div>
             </div>
         </div>
@@ -221,7 +222,7 @@
 </ul>
 </Footer>
 <script src="/client/Javascript/sign-in-sign-out.js"></script>
-<script src="/client/Javascript/movieDetails.js"></script>
+<script type="module" src="Javascript/movieDetails/main.js"></script>
 <script src="/client/Javascript/video-play.js"></script>
